units: update ExecStart path for improved portability
On Fedora 28, /lib is a symlink to /usr/lib. On Ubuntu 18.04 LTS, the two are independent directories.  Furthermore, on Ubuntu 18.04 LTS, all the systemd-* binaries live in /lib/systemd/, not /usr/lib/systemd/.

This changes the ExecStart line to /lib/systemd/systemd-netlogd to improve portability across different Linux distributions.  Closes #11.