Message event loop tweaks
Stop the event source before unregistering it.

Unregister the event loop before closing the journal.

Unset the watch file descriptor when closing the parent journal.

Assert the file descriptor of incoming events has the same number as the
one on creation time (it might however still point to a differnt file).